#75 Refactors view helper functions for credo
