+++ lib/app_web/views/issue_view.ex
@@ -1,22 +1,4 @@
 defmodule AppWeb.IssueView do
   use AppWeb, :view
-
-  def format_date(date) do
-    date
-    |> NaiveDateTime.to_iso8601
-    |> tidyISODate
-  end
-
-  def tidyISODate(iso_date) do
-    iso_date
-    |> String.replace("T", " ")
-    |> String.split(".")
-    |> List.first
-  end
-
-  def display_markdown(text) do
-    text
-    |> Earmark.as_html!
-    |> raw
-  end
+  import AppWeb.Helpers.View
 end
